requirements.txt
fastapi
httpx
python-dotenv
uvicorn
pyyaml
slowapi
mkdocs
mkdocs-material